+++ Converter/Lemur/ConversionManager.cs
@@ -294,6 +294,7 @@ namespace Converter.Lemur
             using (var _ = OperationTimer.Start("Writing mod descriptor")) await ModDescriptorWriter.Write(Settings.Instance.ModName, Settings.Instance.ModsDirectory, Settings.OutputDirectory);
             await LandlessTitleStubsWriter.Write(Settings.OutputDirectory);
             await VanillaEventOverridesWriter.Write(Settings.OutputDirectory);
+            await PoolRepopulationOverrideWriter.Write(Settings.OutputDirectory);
             if (w.MapDefines)
             {
                 using var _ = OperationTimer.Start("Writing map defines");